pyproject.toml
[project]
name = "manga-wellness-backend"
version = "0.1.0"
description = "Manga wellness backend with nano-banana and GCS"
requires-python = ">=3.12"
dependencies = [
    # Core FastAPI dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    # Google AI/Gemini dependencies
    "google-generativeai>=0.3.0",
    "langchain-google-genai>=2.0.0",
    # LangChain
    "langchain>=0.1.0",
    "langchain-core>=0.3.0",
    # Google Cloud services (mandatory)
    "google-cloud-storage>=2.10.0",
    "google-cloud-texttospeech>=2.27.0",
    # Other dependencies
    "pydantic>=2.5.0",
    "pillow>=10.1.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "httpx>=0.25.0",
    "python-socketio>=5.10.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-auth>=2.23.0",
    "langchain-google-vertexai>=2.0.0",
    "pydantic-settings>=2.1.0",
    "langgraph>=0.0.20",
    "streamlit>=1.28.0",
    "aiohttp>=3.9.0",
    "protobuf>=3.20.2,<6.0.0",
    "numpy>=1.21.0",
    "typing-extensions>=4.0.0",
    "pyaudio>=0.2.11",
    # Auth & DB
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = [
    "services",
    "routers", 
    "config",
    "models",
    "utils",
    "workflows"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
